feat: starting avaliacao class business rule

+++ MajorBeatAPI/src/main/java/br/com/harmoniar/MajorBeatAPI/entity/Avaliacao.java
@@ -17,7 +17,7 @@ public class Avaliacao {
     private Long idAvaliacao;
 
     @Column
-    private int nota;
+    private Long nota;
 
     @Column
     private String comentario;
